+++ purldb/urls.py
@@ -8,7 +8,7 @@
 #
 
 from django.conf.urls import include
-from django.urls import re_path
+from django.urls import path
 from django.views.generic import RedirectView
 from rest_framework import routers
 
@@ -23,16 +23,16 @@ from minecode.api import PriorityResourceURIViewSet
 
 
 api_router = routers.DefaultRouter()
-api_router.register(r'packages', PackageViewSet)
-api_router.register(r'resources', ResourceViewSet)
-api_router.register(r'approximate_directory_content_index', ApproximateDirectoryContentIndexViewSet)
-api_router.register(r'approximate_directory_structure_index', ApproximateDirectoryStructureIndexViewSet)
-api_router.register(r'exact_file_index', ExactFileIndexViewSet)
-api_router.register(r'exact_package_archive_index', ExactPackageArchiveIndexViewSet)
-api_router.register(r'cditems', CDitemViewSet, 'cditems')
-api_router.register(r'on_demand_queue', PriorityResourceURIViewSet)
+api_router.register('packages', PackageViewSet)
+api_router.register('resources', ResourceViewSet)
+api_router.register('approximate_directory_content_index', ApproximateDirectoryContentIndexViewSet)
+api_router.register('approximate_directory_structure_index', ApproximateDirectoryStructureIndexViewSet)
+api_router.register('exact_file_index', ExactFileIndexViewSet)
+api_router.register('exact_package_archive_index', ExactPackageArchiveIndexViewSet)
+api_router.register('cditems', CDitemViewSet, 'cditems')
+api_router.register('on_demand_queue', PriorityResourceURIViewSet)
 
 urlpatterns = [
-    re_path(r'^api/', include((api_router.urls, 'api'))),
-    re_path("", RedirectView.as_view(url="api/")),
+    path('api/', include((api_router.urls, 'api'))),
+    path("", RedirectView.as_view(url="api/")),
 ]